Move central store thread to utils
