+++ tendrl/performance_monitoring/sds/glusterfs/glusterfs_plugin.py
@@ -7,8 +7,11 @@ from tendrl.performance_monitoring import constants as \
 from tendrl.performance_monitoring.objects.system_summary \
     import SystemSummary
 from tendrl.performance_monitoring.sds import SDSPlugin
-from tendrl.performance_monitoring.utils import parse_resource_alerts
-from tendrl.performance_monitoring.utils import read as etcd_read_key
+import tendrl.performance_monitoring.utils.central_store_util \
+    as central_store_util
+from tendrl.performance_monitoring.utils.util import parse_resource_alerts
+from tendrl.performance_monitoring.utils.central_store_util \
+    import read as etcd_read_key
 
 
 class GlusterFSPlugin(SDSPlugin):
@@ -26,7 +29,7 @@ class GlusterFSPlugin(SDSPlugin):
     def configure_monitoring(self, sds_tendrl_context):
         configs = []
         cluster_node_ids = \
-            NS.central_store_thread.get_cluster_node_ids(
+            central_store_util.get_cluster_node_ids(
                 sds_tendrl_context['integration_id']
             )
         for node_id in cluster_node_ids:
@@ -443,7 +446,7 @@ class GlusterFSPlugin(SDSPlugin):
             )
 
     def get_node_brick_status_counts(self, node_id):
-        node_name = NS.central_store_thread.get_node_name_from_id(node_id)
+        node_name = central_store_util.get_node_name_from_id(node_id)
         brick_status_wise_counts = {
             'stopped': 0,
             'total': 0,
@@ -451,7 +454,7 @@ class GlusterFSPlugin(SDSPlugin):
             pm_consts.CRITICAL_ALERTS: 0
         }
         try:
-            cluster_id = NS.central_store_thread.get_node_cluster_id(
+            cluster_id = central_store_util.get_node_cluster_id(
                 node_id
             )
             if cluster_id:
